Enable type deduction for `trapping_add`.

+++ trapping_test.cc
@@ -94,17 +94,17 @@ void TestBasicAssumptions() {
 
 void TestAdd() {
   EXPECT_DEATH((trapping_add<i32, i32, i32>(i32_max, 1)));
-  EXPECT_DEATH((trapping_add<i32, i32, i16>(i32_max, 0)));
+  EXPECT_DEATH((trapping_add<i16, i32, i32>(i32_max, 0)));
   EXPECT_DEATH((trapping_add<u32, u32, u32>(u32_max, 1)));
-  EXPECT_DEATH((trapping_add<u32, u32, u16>(u32_max, 0)));
+  EXPECT_DEATH((trapping_add<u16, u32, u32>(u32_max, 0)));
   EXPECT_DEATH((trapping_add<i16, i16, i16>(i16_max, 1)));
-  EXPECT_DEATH((trapping_add<i16, i16, i8>(i16_max, 0)));
+  EXPECT_DEATH((trapping_add<i8, i16, i16>(i16_max, 0)));
   EXPECT_DEATH((trapping_add<u16, u16, u16>(u16_max, 1)));
-  EXPECT_DEATH((trapping_add<u16, u16, u8>(u16_max, 0)));
+  EXPECT_DEATH((trapping_add<u8, u16, u16>(u16_max, 0)));
 
   const i64 expected = trapping_cast<i64>(u32_max) + 1;
-  EXPECT(expected == (trapping_add<u32, u32, i64>(u32_max, 1)));
-  EXPECT(expected == (trapping_add<u32, u16, i64>(u32_max, 1)));
+  EXPECT(expected == (trapping_add<i64, u32, u32>(u32_max, 1)));
+  EXPECT(expected == (trapping_add<i64, u32, u16>(u32_max, 1)));
 }
 
 void TestCastTruncate() {
